+++ pkg/querier/tripperware/queryrange/limits_test.go
@@ -236,7 +236,7 @@ type mockLimits struct {
 	maxQueryLookback       time.Duration
 	maxQueryLength         time.Duration
 	maxCacheFreshness      time.Duration
-	maxQueryResponseSize   int
+	maxQueryResponseSize   int64
 	queryVerticalShardSize int
 }
 
@@ -256,7 +256,7 @@ func (m mockLimits) MaxCacheFreshness(string) time.Duration {
 	return m.maxCacheFreshness
 }
 
-func (m mockLimits) MaxQueryResponseSize(string) int {
+func (m mockLimits) MaxQueryResponseSize(string) int64 {
 	return m.maxQueryResponseSize
 }
