+++ pkg/querier/tripperware/test_shard_by_query_utils.go
@@ -476,7 +476,7 @@ type mockLimits struct {
 	maxQueryLookback     time.Duration
 	maxQueryLength       time.Duration
 	maxCacheFreshness    time.Duration
-	maxQueryResponseSize int
+	maxQueryResponseSize int64
 	shardSize            int
 	queryPriority        validation.QueryPriority
 	queryRejection       validation.QueryRejection
@@ -498,7 +498,7 @@ func (m mockLimits) MaxCacheFreshness(string) time.Duration {
 	return m.maxCacheFreshness
 }
 
-func (m mockLimits) MaxQueryResponseSize(string) int {
+func (m mockLimits) MaxQueryResponseSize(string) int64 {
 	return m.maxQueryResponseSize
 }
